refactor(file-router): small improvements according to reviews
* Make extended ViewConfig type private
* Make Vaadin object type private
* Fix building issue
